Updates which add in pos and neg weights
More Updates. I think that I should probably make it so that the mappings map the full 32x32 array, even if nothing is happening...

At the very least, I need to map to make everyone have at least 32 stuff and 1 bias. Otherwise, it will not make any sense I think?

Maybe that is the compromise that I make. We use 32x1 with bias for everyone, and then go from there. At least at that point, it is the same architecture as the things that we model.